Move iterator code to separate module
